requirements.txt
asgiref==3.8.1
boto3==1.35.98
botocore==1.35.98
Django==5.1.4
django-environ==0.11.2
django-storages==1.14.4
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
jmespath==1.0.1
mysqlclient==2.2.6
packaging==24.2
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
Pillow==9.5.0
